Examples warnings, and Altera container build search fix tools/cdk/examples/bestfit_test/impl2.rcc/impl2.c: size_t warning tools/cdk/examples/bestfit_test/impl3.rcc/impl3.c: size_t warning tools/cdk/examples/ext2file_xml/ext2file.cxx: size_t warning tools/cdk/include/hdl/quartus.mk: workers in the container (not app) were not being found

+++ tools/cdk/examples/bestfit_test/impl2.rcc/impl2.c
@@ -16,7 +16,7 @@ typedef struct {
   double wave[S_SIZE];
 } MyState;
 
-static uint32_t sizes[] = {sizeof(MyState), 0};
+static size_t sizes[] = {sizeof(MyState), 0};
 #define PI 3.14159265
 
 IMPL2_METHOD_DECLARATIONS;
